Revert "redux integration"
