+++ .gitignore
@@ -1,6 +1,2 @@
 /node_modules
 package-lock.json
-src/package.json
-src/store/index.js
-src/yarn.lock
-yarn.lock
