+++ client/.gitignore
@@ -19,4 +19,3 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-
