+++ client/src/App.js
@@ -1,24 +1,20 @@
-import React, { Component } from 'react'
-import { Provider } from 'react-redux'
-import logo from './logo.svg'
-import './App.css'
-import store from './store'
-import Customers from './components/Customer/customers'
+import React, { Component } from 'react';
+import logo from './logo.svg';
+import './App.css';
+import Customers from './components/customers';
 
 class App extends Component {
-  render () {
+  render() {
     return (
-      <Provider store={ store }>
-        <div className="App">
-          <header className="App-header">
-            <img src={ logo } className="App-logo" alt="logo"/>
-            <h1 className="App-title">React Express Starter</h1>
-          </header>
-          <Customers/>
-        </div>
-      </Provider>
-    )
+      <div className="App">
+        <header className="App-header">
+          <img src={logo} className="App-logo" alt="logo" />
+          <h1 className="App-title">React Express Starter</h1>
+        </header>
+        <Customers />
+      </div>
+    );
   }
 }
 
-export default App
+export default App;
