--- client/src/package.json
@@ -1,8 +1,0 @@
-{
-  "dependencies": {
-    "react": "^16.2.0",
-    "react-redux": "^5.0.6",
-    "redux": "^3.7.2",
-    "redux-thunk": "^2.2.0"
-  }
-}
