--- client/src/store/actions/constants.js
@@ -1,2 +1,0 @@
-export const GET_CUSTOMERS = "GET_CUSTOMERS";
-export const UPDATE_CUSTOMER = "REMOVE_CUSTOMER";
