--- client/src/store/actions/customer.js
@@ -1,7 +1,0 @@
-import {GET_CUSTOMERS} from './constants';
-
-export const getCustomers = () => dispatch => {
-  return fetch('/api/customers')
-    .then(res => res.json())
-    .then(customers => dispatch({type: GET_CUSTOMERS, payload: customers}))
-}
