--- client/src/store/reducers/index.js
@@ -1,6 +1,0 @@
-import {combineReducers} from 'redux';
-import customerReducer from './customer';
-
-export default combineReducers({
-  customers: customerReducer
-})
